requirements.txt
gpiod==2.2.2
gpiodevice==0.0.5
inky==2.0.0
numpy==2.1.3
pillow==11.0.0
smbus2==0.5.0
spidev==3.6
